Flatten swap crate structure and move test into `tests`
