+++ swap/Cargo.toml
@@ -7,11 +7,11 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 elements-fun = { path = "../elements-fun", features = [ "serde" ] }
-elements-harness = { path = "../elements-harness" }
-secp256k1 = { version = "0.19.0", features = [ "global-context", "bitcoin_hashes" ] }
+secp256k1 = { version = "0.19.0", features = [ "bitcoin_hashes" ] }
 serde = { version = "1", features = [ "derive" ] }
 sha2 = "0.9"
 
 [dev-dependencies]
+elements-harness = { path = "../elements-harness" }
 testcontainers = "0.11"
 tokio = { version = "0.2", default-features = false, features = [ "blocking", "macros", "rt-core", "time" ] }
